pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "asyncio>=3.4.3",
    "dnspython>=2.7.0",
    "flask-dance>=7.1.0",
    "flask-login>=0.6.3",
    "oauthlib>=3.3.1",
    "openpyxl>=3.1.5",
    "pandas>=2.3.1",
    "pyjwt>=2.10.1",
    "python-telegram-bot>=22.3",
    "sqlalchemy>=2.0.42",
]
